Prefix-based TempURLs
Prefix-based temporary URLs allow temporary access to swift clusters for
users without credentials. This commit adds authentication for temporary
URLs whose prefix is empty (giving access to the whole container).

Instead of the swiftclient, the requests library is now used for
interacting with the server, because the swiftclient cannot handle
temporary URLs in general at this moment. But the swiftclient is still
used for authentication with username/password. With this,
keystone support should still be possible.